requirements.txt
# Core Dependencies
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
pydantic==2.5.0

# AI Providers
openai==1.3.9
google-generativeai==0.3.2
ollama==0.1.7
requests==2.31.0

# TON Blockchain
pytonlib==0.0.14
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
pytz==2023.3
asyncio==3.4.3
aiohttp==3.9.1
